requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
google-generativeai
faiss-cpu
python-dotenv
PyPDF2